[rendering][GLES] Move the HDR GUI composite out of the GBM winsystem
The HDR GUI composite (GUI rendered in sRGB into an FBO, then composited
through the sRGB->PQ/HLG shader into the back buffer) lived in
CWinSystemGbmGLESContext, although the only GBM knowledge in it was three
reads of the DRM plane state answering one question: is the video on a
separate plane, so that something else blends the GUI over it.

Move the bodies of SetGuiCompositing, BeginGuiComposite, EndGuiComposite
and CompositeGui, with their seven members, into a new CGuiCompositeGLES
(Enable, Begin, End, Composite, IsActive). The platform reads become
parameters: videoOnSeparatePlane for the skip-clear, the cached-frame
early return and the blend function choice; limitedRange, the FBO size,
the depth attachment and the GUI element count are passed by the caller.

CRenderSystemGLES owns the instance, so every GLES windowing system gets
one without declaring it. The composite draws with GL into the render
system's back buffer and its state already steers the render system's
GUI shader defines through IsHdrComposite; the windowing system only
supplies the plane topology. DestroyRenderSystem releases the
composite's FBO and shader while the context is still current, because
the Android windowing destroys its EGL context before the
CRenderSystemGLES base and this member. The GBM windowing keeps the five
CWinSystemBase overrides as forwards that compute videoOnSeparatePlane
from the DRM planes exactly as before, so GBM behaviour is unchanged.
The GL variant, CWinSystemGbmGLContext, keeps its own copy.
