+++ xbmc/rendering/gles/RenderSystemGLES.cpp
@@ -165,6 +165,8 @@ bool CRenderSystemGLES::DestroyRenderSystem()
   glFinish();
   PresentRenderImpl(true);
 
+  // release the GUI composite FBO and shader while the context is current
+  m_guiComposite.Enable(0, false);
   ReleaseShaders();
   m_bRenderCreated = false;
