+++ xbmc/windowing/gbm/WinSystemGbmGLESContext.h
@@ -9,13 +9,10 @@
 #pragma once
 
 #include "WinSystemGbmEGLContext.h"
-#include "cores/VideoPlayer/VideoRenderers/FrameBufferObject.h"
 #include "rendering/gles/RenderSystemGLES.h"
 #include "utils/EGLUtils.h"
 
 #include <memory>
-
-class CGuiCompositeShaderGLES;
 
 class CVaapiProxy;
 
@@ -47,7 +44,7 @@ public:
   bool BeginGuiComposite(bool guiWillRender) override;
   void EndGuiComposite() override;
   void CompositeGui() override;
-  bool IsHdrComposite() const override { return m_guiCompositing; }
+  bool IsHdrComposite() const override { return m_guiComposite.IsActive(); }
 
 protected:
   void SetVSyncImpl(bool enable) override {}
@@ -55,16 +52,7 @@ protected:
   bool CreateContext() override;
 
 private:
-  bool m_guiCompositing{false};
-  CFrameBufferObject m_guiFbo;
-  int m_guiFboWidth{0};
-  int m_guiFboHeight{0};
-  // True when the GUI FBO is empty (no draws this frame); CompositeGui skips composite when true.
-  bool m_guiFboClean{false};
-  // Whether the GUI render pass will run this frame; set by BeginGuiComposite.
-  bool m_guiWillRender{true};
-
-  std::unique_ptr<CGuiCompositeShaderGLES> m_compositeShader;
+  bool VideoOnSeparatePlane() const;
 };
 
 } // namespace GBM
